Move the substrate keystore proto define to substrate
